Removed a whole lot of clutter Added a "statusCallback" parameter to the client to allow for custom status updates and processing on the fly
A couple of thoughts - it might make sense to allow for better processing in "real time" using the callback instead of writing everything to temp files.  The "problem" is that the results don't necessarily come back in the order that they went out.  But a temporary queue could be setup to allow for that.  More to explore.